[refactor] index -> createStatementData/statement로  파일 분리
