+++ cp1/index.html
@@ -7,6 +7,6 @@
   <title>Document</title>
 </head>
 <body>
-  <script type="module" src="./src/index.js"></script>
+  <script type="module" src="./src/statement.js"></script>
 </body>
 </html>
