+++ cp1/src/statement.js
@@ -1,5 +1,6 @@
 import inovices from "./inovices.js";
 import plays from "./plays.js";
+import createStatementData from "./createStatementData.js";
 
 function statement(inovice, plays) {
   const statementData = {};
@@ -7,7 +8,7 @@ function statement(inovice, plays) {
   statementData.performances = inovice.performances.map(enrichPerformance);
   statementData.totalAmount = totalAmount(statementData);
   statementData.totalVolumeCredits = totalVolumeCredits(statementData);
-  return renderPlainText(statementData, plays);
+  return createStatementData(statementData, plays);
 
   function enrichPerformance(aPerformance) {
     const result = Object.assign({}, aPerformance);
@@ -66,26 +67,6 @@ function statement(inovice, plays) {
       result += perf.volumeCredits;
     }
     return result;
-  }
-}
-function renderPlainText(data, plays) {
-  let result = `청구 내역 (고객명: ${data.customer})\n`;
-
-  for (let perf of data.performances) {
-    //청구 내역을 출력한다.
-    result += ` ${perf.play.name}: ${usd(perf.amount)} (${perf.audience}석)\n`;
-  }
-
-  result += `총액: ${usd(data.totalAmount)}\n`;
-  result += `적립 포인트: ${data.totalVolumeCredits}점\n`;
-  return result;
-
-  function usd(aNumber) {
-    return new Intl.NumberFormat("en-US", {
-      style: "currency",
-      currency: "USD",
-      minimumFractionDigits: 2,
-    }).format(aNumber / 100);
   }
 }
